refactor(ui): extract LearningsDrawer leaf cards + LearningsCtx (#855)
Creates learnings-drawer/ subdir with ctx.ts + 4 leaf-card components:
InjectableRuleCard, ProposedRuleCard, MergeSuggestCard, RetiredRuleCard.
Shared scope-editor state threaded via $derived LearningsCtx to preserve
single-open semantics. Per-card local state (draft, expanded) moved to
each card. Matching CSS migrated per plan's style map.

Parent template: cyclomatic 63→40 / cognitive 157→94 (still over bar;
later tasks finish the split).

[no-feature-entry]
