+++ ui/src/lib/components/LearningsDrawer.svelte
@@ -1,27 +1,17 @@
 <script lang="ts">
   import { fly } from "svelte/transition";
-  import type { Action } from "svelte/action";
   import { dialog } from "$lib/a11yDialog";
   import { m } from "$lib/paraglide/messages";
-  import type {
-    InjectableRule,
-    Learning,
-    RepoInjectable,
-    MergeSuggestion,
-    SignalKind,
-  } from "$lib/types";
+  import type { InjectableRule, Learning, RepoInjectable, MergeSuggestion } from "$lib/types";
   import { learnings } from "$lib/learnings.svelte";
   import { projectIcons } from "$lib/projectIcons.svelte";
   import {
     basename,
     repoAnchorId,
     mergeRepoGroups,
-    injectionBadge,
     injectedCount,
-    showIneffective,
     flaggedCount,
     totalFlagged,
-    evidenceSources,
     sortGroupsForTriage,
     isOverBudget,
     droppedCount,
@@ -31,50 +21,12 @@
     retiredRules,
     retiredCount,
     unseenRetiredCount,
-    helpRate,
   } from "./learnings-drawer";
-
-  // Human label for an evidence source kind. Mirrors the server SignalKind set;
-  // keeps the i18n call in the component (the helper stays locale-agnostic).
-  const kindLabel = (k: SignalKind): string =>
-    k === "reply"
-      ? m.learnings_kind_reply()
-      : k === "critic"
-        ? m.learnings_kind_critic()
-        : k === "block"
-          ? m.learnings_kind_block()
-          : m.learnings_kind_stall();
-
-  // Grow a rule textarea to fit its content so the full rule is always readable —
-  // no inner scroll, no clipping. Re-measures on every input and, via `update`,
-  // whenever the bound value changes programmatically (e.g. a poll refresh
-  // replacing the rule text while no local draft exists).
-  const autosize: Action<HTMLTextAreaElement, string> = (el) => {
-    let raf = 0;
-    const fit = () => {
-      el.style.height = "auto";
-      el.style.height = `${el.scrollHeight}px`;
-    };
-    // scrollHeight is unreliable mid fly-in / before the new value is flushed
-    // to the DOM; settle one frame later.
-    const fitNextFrame = () => {
-      cancelAnimationFrame(raf);
-      raf = requestAnimationFrame(fit);
-    };
-    fit();
-    fitNextFrame();
-    el.addEventListener("input", fit);
-    return {
-      update: fitNextFrame,
-      destroy() {
-        cancelAnimationFrame(raf);
-        el.removeEventListener("input", fit);
-      },
-    };
-  };
-
-  // Which rules have their evidence-source list expanded, keyed by learning id.
-  let expanded = $state<Record<string, boolean>>({});
+  import type { LearningsCtx } from "./learnings-drawer/ctx";
+  import InjectableRuleCard from "./learnings-drawer/InjectableRuleCard.svelte";
+  import ProposedRuleCard from "./learnings-drawer/ProposedRuleCard.svelte";
+  import MergeSuggestCard from "./learnings-drawer/MergeSuggestCard.svelte";
+  import RetiredRuleCard from "./learnings-drawer/RetiredRuleCard.svelte";
 
   // "What is this?" explainer at the top of the drawer. Default open so a first-time
   // user gets the plain-language explanation; collapsing is remembered (localStorage)
@@ -147,14 +99,11 @@
   // Cross-repo card pending the inline two-step confirm before a global CLAUDE.md write (#872).
   let confirmingGlobalId = $state<string | null>(null);
 
-  let drafts = $state<Record<string, string>>({});
-  const draft = (l: Learning) => drafts[l.id] ?? l.rule;
-
   // Glob-scope editing (#842): which injectable rule has its scope editor open, and the
   // comma-separated draft text. Saving splits on commas/whitespace into a glob array.
   let editingScope = $state<string | null>(null);
   let scopeDraft = $state("");
-  function openScopeEditor(l: Learning & { scopeGlobs: string[] }) {
+  function openScopeEditor(l: InjectableRule) {
     editingScope = l.id;
     scopeDraft = l.scopeGlobs.join(", ");
   }
@@ -166,6 +115,30 @@
     onscope(id, globs);
     editingScope = null;
   }
+
+  // Shared context bundle — built as $derived so children's reads of
+  // editingScope/scopeDraft stay reactive (mirroring Herd.svelte:252 rowCtx pattern).
+  const ctx = $derived<LearningsCtx>({
+    onapprove,
+    ondismiss,
+    ondistill,
+    onpromote,
+    onoptimize,
+    onoptimizeall,
+    onrestore,
+    onscope,
+    onseenretired,
+    onmerge,
+    ondismissmerge,
+    onpromoteglobal,
+    onmergenow,
+    editingScope,
+    scopeDraft,
+    onScopeOpen: openScopeEditor,
+    onScopeInput: (v: string) => (scopeDraft = v),
+    onScopeSave: saveScope,
+    onScopeCancel: () => (editingScope = null),
+  });
 
   const reduceMotion =
     typeof window !== "undefined" &&
@@ -388,102 +361,6 @@
     </section>
   {/if}
 
-  <!-- Change 6: Injectable-rule snippet — single source, no copy-paste.
-       Accepts the repo's enabled flag so the badge reflects injection state correctly. -->
-  {#snippet irule(r: InjectableRule, enabled: boolean)}
-    {@const badge = injectionBadge(r, enabled)}
-    <article class="irule">
-      <p class="itext">{r.rule}</p>
-      <div class="ifoot">
-        <span class="chip" class:promoted={r.status === "promoted"}>
-          {r.status === "promoted" ? m.learnings_status_promoted() : m.learnings_status_active()}
-        </span>
-        {#if badge === "injected"}
-          <span class="badge ok">✓ {m.learnings_injected_badge()}</span>
-        {:else if badge === "scoped"}
-          <span class="badge scoped" title={m.learnings_scoped_title()}>
-            ◎ {m.learnings_scoped_badge()}
-          </span>
-        {:else if badge === "over-budget"}
-          <span class="badge warn" title={m.learnings_overbudget_title()}>
-            ⊘ {m.learnings_overbudget_badge()}
-          </span>
-        {:else}
-          <span class="badge off">⊘ {m.learnings_injection_disabled_badge()}</span>
-        {/if}
-        {#if showIneffective(r)}
-          <span class="badge bad" title={m.learnings_ineffective_title()}>
-            ⚠ {m.learnings_ineffective_badge({ count: r.ineffectiveCount })}
-          </span>
-        {/if}
-        {#if helpRate(r) !== null}
-          {@const hr = helpRate(r)!}
-          <span class="help-rate"
-            >{m.learnings_help_rate({ helped: hr.helped, pulls: hr.pulls })}</span
-          >
-        {/if}
-        <span class="spacer"></span>
-        <div class="iactions">
-          {#if showIneffective(r)}
-            <button
-              class="optimize"
-              type="button"
-              onclick={() => onoptimize(r.id)}
-              aria-label={m.learnings_optimize_aria()}
-            >
-              {m.learnings_optimize()}
-            </button>
-          {/if}
-          {#if r.status === "active"}
-            <!-- Change 9: de-emphasised Dismiss with margin-left gap from Promote -->
-            <button class="dismiss dismiss-muted" onclick={() => ondismiss(r.id)}>
-              {m.learnings_dismiss()}
-            </button>
-            <button
-              class="promote"
-              onclick={() => onpromote(r.id)}
-              aria-label={m.learnings_promote_aria()}
-            >
-              {m.learnings_promote()}
-            </button>
-          {:else if r.status === "promoted" && r.promotedPrUrl}
-            <a class="prlink" href={r.promotedPrUrl} target="_blank" rel="noopener external">
-              {m.learnings_promoted_pr()}
-            </a>
-          {/if}
-        </div>
-      </div>
-      <!-- #842: glob scope — which files this rule applies to (empty = always). -->
-      <div class="iscope">
-        {#if editingScope === r.id}
-          <input
-            class="scope-input"
-            type="text"
-            bind:value={scopeDraft}
-            placeholder={m.learnings_scope_placeholder()}
-            aria-label={m.learnings_scope_input_aria()}
-          />
-          <button class="scope-save" type="button" onclick={() => saveScope(r.id)}>
-            {m.common_save()}
-          </button>
-          <button class="scope-cancel" type="button" onclick={() => (editingScope = null)}>
-            {m.common_cancel()}
-          </button>
-        {:else}
-          <span class="scope-label">{m.learnings_scope_label()}</span>
-          {#if r.scopeGlobs.length > 0}
-            {#each r.scopeGlobs as g (g)}<code class="scope-glob">{g}</code>{/each}
-          {:else}
-            <span class="scope-always">{m.learnings_scope_always()}</span>
-          {/if}
-          <button class="scope-edit" type="button" onclick={() => openScopeEditor(r)}>
-            {m.learnings_scope_edit()}
-          </button>
-        {/if}
-      </div>
-    </article>
-  {/snippet}
-
   {#if empty}
     <p class="empty">{m.learnings_empty()}</p>
   {:else}
@@ -532,70 +409,7 @@
 
         <!-- Change 7: hide proposals under either active lens -->
         {#each flaggedOnly || overBudgetOnly ? [] : group.proposed as l (l.id)}
-          <article class="rule">
-            <textarea
-              class="text"
-              rows="2"
-              data-1p-ignore
-              use:autosize={draft(l)}
-              value={draft(l)}
-              oninput={(e) => (drafts = { ...drafts, [l.id]: e.currentTarget.value })}
-              aria-label={m.learnings_rule_aria()}></textarea>
-            {#if l.rationale}
-              <p class="why"><span>{m.learnings_rationale_label()}:</span> {l.rationale}</p>
-            {/if}
-            <div class="evidence" title={m.learnings_evidence_help()}>
-              <span class="evi">{m.learnings_evidence({ count: l.evidenceCount })}</span>
-              {#each evidenceSources(l) as src (src.kind)}
-                <span class="src">{src.count}× {kindLabel(src.kind)}</span>
-              {/each}
-              <!-- render whenever the payload carries provenance (even an empty,
-                   fully-pruned list): the toggle is the focusable route to the
-                   help text for keyboard/touch users -->
-              {#if l.evidenceDetail}
-                <button
-                  class="sources-toggle"
-                  type="button"
-                  title={m.learnings_evidence_help()}
-                  aria-expanded={!!expanded[l.id]}
-                  aria-controls="sources-{l.id}"
-                  aria-label={m.learnings_sources_toggle_aria()}
-                  onclick={() => (expanded = { ...expanded, [l.id]: !expanded[l.id] })}
-                >
-                  {m.learnings_sources_toggle()}
-                  <span class="caret" class:open={expanded[l.id]} aria-hidden="true">▸</span>
-                </button>
-              {/if}
-            </div>
-            {#if expanded[l.id] && l.evidenceDetail}
-              <!-- disclosed region the toggle's aria-controls points at; carries a
-                   visible copy of the provenance explanation, reachable for
-                   keyboard/touch users who can't hover the title tooltip -->
-              <div class="sources-region" id="sources-{l.id}">
-                <p class="shelp">{m.learnings_evidence_help()}</p>
-                {#if l.evidenceDetail.length > 0}
-                  <ul class="sources">
-                    {#each l.evidenceDetail as ev (ev.id)}
-                      <li class="source">
-                        <span class="src">{kindLabel(ev.kind)}</span>
-                        <span class="desig">{ev.desig ?? m.learnings_source_unknown()}</span>
-                        <span class="excerpt">{ev.excerpt}</span>
-                      </li>
-                    {/each}
-                  </ul>
-                {/if}
-              </div>
-            {/if}
-            <div class="foot">
-              <span class="spacer"></span>
-              <button class="dismiss" onclick={() => ondismiss(l.id)}
-                >{m.learnings_dismiss()}</button
-              >
-              <button class="approve" onclick={() => onapprove(l.id, draft(l))}>
-                {m.learnings_approve()}
-              </button>
-            </div>
-          </article>
+          <ProposedRuleCard learning={l} {ctx} />
         {/each}
 
         <!-- Phase 4: intra-repo merge-group suggestions (hidden under the active lenses,
@@ -604,25 +418,7 @@
           <div class="merge-suggest">
             <span class="ms-title">{m.learnings_merge_heading()}</span>
             {#each intraFor(group.repoPath) as s (s.id)}
-              <article class="ms-card">
-                <ul class="ms-members">
-                  {#each s.members ?? [] as mem (mem.id)}
-                    <li class="ms-member">{mem.rule}</li>
-                  {/each}
-                </ul>
-                <p class="ms-result">
-                  <span class="ms-result-label">{m.learnings_merge_result_label()}</span>
-                  {s.mergedRule}
-                </p>
-                <div class="ms-foot">
-                  <button class="ms-dismiss" type="button" onclick={() => ondismissmerge(s.id)}>
-                    {m.learnings_dismiss()}
-                  </button>
-                  <button class="ms-apply" type="button" onclick={() => onmerge(s.id)}>
-                    {m.learnings_merge_apply()}
-                  </button>
-                </div>
-              </article>
+              <MergeSuggestCard suggestion={s} {ctx} />
             {/each}
           </div>
         {/if}
@@ -658,18 +454,26 @@
                 </button>
               {/if}
             </div>
-            <!-- Change 6: dropped-first rendering via snippet + splitDropped -->
+            <!-- Change 6: dropped-first rendering via component + splitDropped -->
             {#if lensActive}
               {#each visibleInjectableRules(inj, { flaggedOnly, overBudgetOnly }) as r (r.id)}
-                {@render irule(r, inj.enabled)}
+                <InjectableRuleCard rule={r} enabled={inj.enabled} {ctx} />
               {/each}
             {:else}
               {@const split = splitDropped(inj)}
               {#if split.dropped.length > 0}
                 <p class="not-injected-label">{m.learnings_not_injected_label()}</p>
-                {#each split.dropped as r (r.id)}{@render irule(r, inj.enabled)}{/each}
+                {#each split.dropped as r (r.id)}<InjectableRuleCard
+                    rule={r}
+                    enabled={inj.enabled}
+                    {ctx}
+                  />{/each}
               {/if}
-              {#each split.injected as r (r.id)}{@render irule(r, inj.enabled)}{/each}
+              {#each split.injected as r (r.id)}<InjectableRuleCard
+                  rule={r}
+                  enabled={inj.enabled}
+                  {ctx}
+                />{/each}
             {/if}
           </div>
         {/if}
@@ -682,27 +486,7 @@
               >
             </div>
             {#each retiredRules(group.injectable) as r (r.id)}
-              <article class="retired-rule">
-                <p class="retired-text">{r.rule}</p>
-                <p class="retired-reason">
-                  {m.learnings_retired_reason({
-                    helped: r.helpfulCount,
-                    pulls: r.injectedCount,
-                    flagged: r.ineffectiveCount,
-                  })}
-                </p>
-                <div class="retired-foot">
-                  <span class="spacer"></span>
-                  <button
-                    class="restore"
-                    type="button"
-                    aria-label={m.learnings_restore_aria()}
-                    onclick={() => onrestore(r.id)}
-                  >
-                    {m.learnings_restore()}
-                  </button>
-                </div>
-              </article>
+              <RetiredRuleCard rule={r} {ctx} />
             {/each}
           </div>
         {/if}
@@ -935,79 +719,6 @@
     padding: 3px 8px;
     cursor: pointer;
   }
-  .rule {
-    border: 1px solid var(--color-line);
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-  }
-  .text {
-    width: 100%;
-    /* height is driven by the autosize action so the full rule always shows */
-    resize: none;
-    overflow: hidden;
-    background: var(--color-inset);
-    color: var(--color-ink-bright);
-    border: 1px solid var(--color-line);
-    padding: 8px 10px;
-    font: inherit;
-    font-size: var(--fs-base);
-    line-height: 1.5;
-  }
-  .why {
-    font-size: var(--fs-base);
-    color: var(--color-muted);
-    line-height: 1.5;
-  }
-  .why span {
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    font-size: var(--fs-micro);
-  }
-  /* Evidence provenance — its own row so the source breakdown has room to wrap.
-     Hover the row for the help tooltip explaining what a signal is. */
-  .evidence {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: 4px 6px;
-    cursor: help;
-  }
-  .foot {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  }
-  .evi {
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-  }
-  .src {
-    font-size: var(--fs-meta);
-    color: var(--color-ink-bright);
-    background: var(--color-inset);
-    border: 1px solid var(--color-line);
-    border-radius: 3px;
-    padding: 1px 6px;
-    white-space: nowrap;
-  }
-  .sources-toggle {
-    margin-left: auto;
-    display: inline-flex;
-    align-items: center;
-    gap: 3px;
-    background: none;
-    border: none;
-    padding: 0;
-    cursor: pointer;
-    font: inherit;
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-  }
-  .sources-toggle:hover {
-    color: var(--color-ink-bright);
-  }
   .caret {
     display: inline-block;
     transition: transform 0.12s ease;
@@ -1015,83 +726,6 @@
   .caret.open {
     transform: rotate(90deg);
   }
-  /* Evidence provenance: the actual signals this rule was distilled from. */
-  .sources-region {
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-  }
-  .shelp {
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-    line-height: 1.5;
-  }
-  .sources {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-  }
-  .source {
-    display: grid;
-    grid-template-columns: auto auto 1fr;
-    align-items: baseline;
-    gap: 6px;
-    font-size: var(--fs-base);
-    line-height: 1.4;
-  }
-  .desig {
-    color: var(--color-ink-bright);
-    font-weight: 600;
-    white-space: nowrap;
-    font-variant-numeric: tabular-nums;
-  }
-  .excerpt {
-    color: var(--color-muted);
-    min-width: 0;
-  }
-  .spacer {
-    flex: 1;
-  }
-  .dismiss,
-  .approve {
-    font-size: var(--fs-base);
-    padding: 5px 12px;
-    cursor: pointer;
-    border: 1px solid var(--color-line-bright);
-    background: none;
-    color: var(--color-muted);
-  }
-  .approve {
-    border-color: var(--color-green);
-    color: var(--color-green);
-  }
-  .promote {
-    font-size: var(--fs-base);
-    padding: 5px 12px;
-    cursor: pointer;
-    border: 1px solid var(--color-green);
-    background: none;
-    color: var(--color-green);
-  }
-  .prlink {
-    font-size: var(--fs-base);
-    color: var(--color-green);
-    text-decoration: none;
-  }
-  /* Per-rule "Optimize" — headline action for a flagged rule, styled primary-ish
-     like .promote but in the amber "needs attention" hue that flags the rule. */
-  .optimize {
-    font-size: var(--fs-base);
-    padding: 5px 12px;
-    cursor: pointer;
-    border: 1px solid var(--color-amber);
-    background: none;
-    color: var(--color-amber);
-  }
-
   /* ── injected house rules ────────────────────────────────────────────── */
   .injected {
     display: flex;
@@ -1139,128 +773,6 @@
     color: var(--color-amber);
     margin: 4px 0 2px;
   }
-  .irule {
-    border: 1px solid var(--color-line);
-    padding: 8px 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-  }
-  .itext {
-    font-size: var(--fs-base);
-    color: var(--color-ink-bright);
-    line-height: 1.5;
-  }
-  .ifoot {
-    display: flex;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: 6px;
-  }
-  /* Action group stays together as one flex item so it wraps as a unit
-     (never fragmenting Optimize/Dismiss/Promote across lines) when the
-     badge row runs out of room in the narrow drawer. */
-  .iactions {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-  }
-  /* Change 9: de-emphasised Dismiss on active rules — muted + gap from Promote */
-  .dismiss-muted {
-    color: var(--color-muted);
-    border-color: var(--color-line);
-    margin-right: 4px;
-  }
-  .chip {
-    font-size: var(--fs-micro);
-    text-transform: uppercase;
-    letter-spacing: 0.06em;
-    padding: 2px 6px;
-    border: 1px solid var(--color-line-bright);
-    color: var(--color-muted);
-  }
-  .chip.promoted {
-    border-color: var(--color-green);
-    color: var(--color-green);
-  }
-  .badge {
-    font-size: var(--fs-meta);
-    padding: 2px 6px;
-    border: 1px solid var(--color-line);
-  }
-  .badge.ok {
-    border-color: var(--color-green);
-    color: var(--color-green);
-  }
-  .badge.warn {
-    border-color: var(--color-amber);
-    color: var(--color-amber);
-    cursor: help;
-  }
-  .badge.bad {
-    border-color: var(--color-red, var(--color-amber));
-    color: var(--color-red, var(--color-amber));
-    cursor: help;
-  }
-  .badge.off {
-    color: var(--color-muted);
-  }
-  .badge.scoped {
-    border-color: var(--color-blue);
-    color: var(--color-blue);
-    cursor: help;
-  }
-  .iscope {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    gap: 6px;
-    margin-top: 6px;
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-  }
-  .scope-label {
-    color: var(--color-muted);
-  }
-  .scope-glob {
-    padding: 1px 5px;
-    border: 1px solid var(--color-line-bright);
-    color: var(--color-blue);
-    background: color-mix(in srgb, var(--color-blue) 10%, var(--color-head));
-  }
-  .scope-always {
-    color: var(--color-muted);
-    font-style: italic;
-  }
-  .scope-edit,
-  .scope-save,
-  .scope-cancel {
-    font-size: var(--fs-meta);
-    background: none;
-    border: 1px solid var(--color-line-bright);
-    color: var(--color-muted);
-    padding: 2px 7px;
-    cursor: pointer;
-  }
-  .scope-edit:hover,
-  .scope-save:hover,
-  .scope-cancel:hover {
-    color: var(--color-ink-bright);
-    border-color: var(--color-ink-bright);
-  }
-  .scope-input {
-    flex: 1 1 12rem;
-    min-width: 0;
-    font-size: var(--fs-meta);
-    padding: 2px 6px;
-    background: var(--color-inset);
-    border: 1px solid var(--color-line-bright);
-    color: var(--color-ink-bright);
-  }
-  .help-rate {
-    font-size: var(--fs-micro);
-    color: var(--color-muted);
-  }
   .hw-review {
     align-self: flex-start;
     font-size: var(--fs-meta);
@@ -1290,42 +802,6 @@
     letter-spacing: 0.1em;
     color: var(--color-muted);
   }
-  .retired-rule {
-    border: 1px solid var(--color-line);
-    padding: 8px 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-    opacity: 0.7;
-  }
-  .retired-text {
-    font-size: var(--fs-base);
-    color: var(--status-done);
-    line-height: 1.5;
-  }
-  .retired-reason {
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-    line-height: 1.45;
-  }
-  .retired-foot {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  }
-  .restore {
-    font-size: var(--fs-base);
-    padding: 4px 10px;
-    cursor: pointer;
-    border: 1px solid var(--color-line-bright);
-    background: none;
-    color: var(--color-muted);
-  }
-  .restore:hover {
-    color: var(--color-ink-bright);
-    border-color: var(--color-ink-bright);
-  }
-
   /* Phase 4: cross-repo recurrence band (top-level, repo-agnostic) */
   .recur {
     display: flex;
@@ -1390,36 +866,6 @@
     text-transform: uppercase;
     letter-spacing: 0.1em;
     color: var(--color-muted);
-  }
-  .ms-members {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 3px;
-  }
-  .ms-member {
-    font-size: var(--fs-meta);
-    color: var(--color-muted);
-    line-height: 1.45;
-    padding-left: 12px;
-    position: relative;
-  }
-  .ms-member::before {
-    content: "•";
-    position: absolute;
-    left: 2px;
-    color: var(--color-line-bright);
-  }
-  .ms-result {
-    font-size: var(--fs-base);
-    color: var(--color-ink-bright);
-    line-height: 1.5;
-  }
-  .ms-result-label {
-    color: var(--color-muted);
-    font-size: var(--fs-meta);
   }
   /* Merge = actionable consolidation → green (matches .promote) */
   .ms-apply {
